coding style cleanups bug fix with some ATOMIC_ADD that was missused

+++ DEA/AccMult/trunk/tools/dag-dot.c
@@ -23,5 +23,6 @@ void terminate_dat_dot(void)
 
 void add_deps(uint64_t child, uint64_t father)
 {
-	fprintf(out_file, "\t \"%llx\"->\"%llx\"\n", father, child);
+	fprintf(out_file, "\t \"%llx\"->\"%llx\"\n", 
+		(unsigned long long)father, (unsigned long long)child);
 }
